+++ src/encoded/types/file_set.py
@@ -280,9 +280,15 @@ def validate_compatible_assay_and_sequencer_on_add(context, request):
     assays = []
     valid_sequencers = []
     for library in data['libraries']:
-        assay_aid = library_utils.get_assay(
-            get_item_or_none(request, library, 'library')
-        )
+        # FIX: 2024-08-29/dmichaels
+        # Guard against passing None to library_utils.get_assay.
+        # TODO: What should really be done here if assay_aid is None (continue) ??
+        # assay_aid = library_utils.get_assay(
+        #     get_item_or_none(request, library, 'library')
+        # )
+        if (library := get_item_or_none(request, library, 'library')) is None:
+            continue
+        assay_aid = library_utils.get_assay(library)
         assay = get_item_or_none(request, assay_aid, 'assay')
         assays.append(
             item_utils.get_identifier(assay)